# ⚡  Make Cloud Logging mode a logger configuration-time option