--- a/.mutmut-cache
+++ b/.mutmut-cache
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:79c9db6fd92ac1fff97b6402447a5998bf01e87d17688809e5f06e235ba99166
-size 122880
+oid sha256:333dbabcb7ed545ba1a93014f80cbe1c0f679b6f39479cc8d760eae2027fbb65
+size 126976